Rename setUpNotifications -> setupNotifications and setUpQuickActions -> setupQuickActions
