+++ .gitignore
@@ -774,9 +774,9 @@ packages/app-mobile/services/AlarmServiceDriver.android.js.map
 packages/app-mobile/services/AlarmServiceDriver.ios.d.ts
 packages/app-mobile/services/AlarmServiceDriver.ios.js
 packages/app-mobile/services/AlarmServiceDriver.ios.js.map
-packages/app-mobile/setUpQuickActions.d.ts
-packages/app-mobile/setUpQuickActions.js
-packages/app-mobile/setUpQuickActions.js.map
+packages/app-mobile/setupQuickActions.d.ts
+packages/app-mobile/setupQuickActions.js
+packages/app-mobile/setupQuickActions.js.map
 packages/app-mobile/utils/ShareExtension.d.ts
 packages/app-mobile/utils/ShareExtension.js
 packages/app-mobile/utils/ShareExtension.js.map
@@ -792,9 +792,9 @@ packages/app-mobile/utils/checkPermissions.js.map
 packages/app-mobile/utils/fs-driver-rn.d.ts
 packages/app-mobile/utils/fs-driver-rn.js
 packages/app-mobile/utils/fs-driver-rn.js.map
-packages/app-mobile/utils/setUpNotifications.d.ts
-packages/app-mobile/utils/setUpNotifications.js
-packages/app-mobile/utils/setUpNotifications.js.map
+packages/app-mobile/utils/setupNotifications.d.ts
+packages/app-mobile/utils/setupNotifications.js
+packages/app-mobile/utils/setupNotifications.js.map
 packages/app-mobile/utils/shareHandler.d.ts
 packages/app-mobile/utils/shareHandler.js
 packages/app-mobile/utils/shareHandler.js.map
