+++ packages/app-mobile/utils/setupNotifications.ts
